Separating out half-integers
